django/pyproject.toml
[tool.poetry]
name = "docker-django"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.8"
Django = "^3.1.3"
djangorestframework = "^3.12.2"
django-cors-headers = "^3.5.0"
django-extensions = "^3.0.9"
django-filter = "^2.4.0"
gunicorn = "^20.0.4"
pendulum = "^2.1.2"
psycopg2-binary = "^2.8.6"

[tool.poetry.dev-dependencies]
black = "^19.10b0"
flake8 = "^3.7.9"
isort = "^4.3.21"
mypy = "^0.770"
pytest = "^6.0.2"
pytest-cov = "^2.10.1"
pytest-django = "^4.1.0"

[tool.black]
line-length = 79

[tool.isort]
line_length = 79
force_single_line = true

[tool.pytest.ini_options]
# addopts = "--capture=no -x --doctest-modules -v --cov=. --cov-report=html"
addopts = "--capture=sys -x --doctest-modules -v --cov=. --cov-report=html"
cache_dir = "/tmp/.pytest_cache"
python_files = [
    "tests.py",
    "test_*.py",
]
python_classes = "Test"
python_functions = "test_"
testpaths = "."

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
